agents/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
google-analytics-data>=0.18.0
google-auth>=2.34.0
